[Clean] Clean codes (in Editor)

+++ Engine/Editor/include/Editor/EditorEngine.h
@@ -20,7 +20,7 @@ namespace Editor
 		~EditorEngine();
 
 		int		Init(const int width, const int height) override;
-		void	SetSizeWindow(const double width, const double height) override;
+		void	SetSizeWindow(const int width, const int height) override;
 		void	Update() override;
 
 		virtual Core::Maths::Vec2	GetMousePos() noexcept override;
